Rewriting and expanding main function as a test.
Adds `ToSlice` helper to the GroupOfCells struct to turn the linked
list back into a []uint8.

Pulls in a third-party testing library just to demonstrate using
one. The changes to `go.mod` and `go.sum` happened automatically as a
result of adding `github.com/stretchr/testify` to the imports and
building the project.